third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-lifetime-reload-expected.txt

Test reloading during encrypted media playback.

EVENT(encrypted)
EVENT(encrypted)
EVENT(message)
EVENT(playing)
This is a testharness.js-based test.
PASS Reloading during encrypted media playback. 
Harness: the test ran to completion.
